some javadoc rewrites and update broadcast receiver related structure
