+++ packages/connectivity/android/src/main/java/io/flutter/plugins/connectivity/ConnectivityBroadcastReceiver.java
@@ -1,0 +1,48 @@
+package io.flutter.plugins.connectivity;
+
+import android.content.BroadcastReceiver;
+import android.content.Context;
+import android.content.Intent;
+import android.content.IntentFilter;
+import android.net.ConnectivityManager;
+import androidx.annotation.NonNull;
+import io.flutter.plugin.common.EventChannel;
+
+/**
+ * The ConnectivityBroadcastReceiver receives the connectivity updates and send them to the UIThread
+ * through an {@link EventChannel.EventSink}
+ *
+ * <p>Use {@link
+ * io.flutter.plugin.common.EventChannel#setStreamHandler(io.flutter.plugin.common.EventChannel.StreamHandler)}
+ * to set up the receiver.
+ */
+public class ConnectivityBroadcastReceiver extends BroadcastReceiver
+    implements EventChannel.StreamHandler {
+  private Context context;
+  private Connectivity connectivity;
+  private EventChannel.EventSink events;
+
+  public ConnectivityBroadcastReceiver(
+      @NonNull Context context, @NonNull Connectivity connectivity) {
+    this.context = context;
+    this.connectivity = connectivity;
+  }
+
+  @Override
+  public void onListen(Object arguments, EventChannel.EventSink events) {
+    this.events = events;
+    context.registerReceiver(this, new IntentFilter(ConnectivityManager.CONNECTIVITY_ACTION));
+  }
+
+  @Override
+  public void onCancel(Object arguments) {
+    context.unregisterReceiver(this);
+  }
+
+  @Override
+  public void onReceive(Context context, Intent intent) {
+    if (events != null) {
+      events.success(connectivity.checkNetworkType());
+    }
+  }
+}
